// apps/web/generated/api/statisticPokemon.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Poketinder
 * OpenAPI spec version: 1.0
 */

export interface StatisticPokemon {
  /** Уникальный идентификатор статистики */
  id: string;
  /** Дизлайки */
  pass: number;
  /** Идентификатор покемона */
  pokemonId: number;
  /** Лайки */
  smash: number;
}